replace older falcon.c by the new version
